+++ src/main/kotlin/com/github/sharoha/cpjcodegen/actions/contest/LeetcodeContestProblemAction.kt
@@ -1,13 +1,14 @@
-package com.github.sharoha.cpjcodegen.actions
+package com.github.sharoha.cpjcodegen.actions.contest
 
+import com.github.sharoha.cpjcodegen.actions.AddProblemAction
 import com.intellij.openapi.project.Project
 
 class LeetcodeContestProblemAction: AddProblemAction<LeetcodeContestDialogWrapper>() {
     override fun getDialog(project: Project): LeetcodeContestDialogWrapper {
-        TODO("Not yet implemented")
+        return LeetcodeContestDialogWrapper(project)
     }
 
     override fun actionComplete() {
-        TODO("Not yet implemented")
+        println("Dialog process successfully completed")
     }
 }
