+++ src/main/kotlin/com/github/sharoha/cpjcodegen/actions/problem/LeetcodeProblemAction.kt
@@ -1,10 +1,11 @@
-package com.github.sharoha.cpjcodegen.actions
+package com.github.sharoha.cpjcodegen.actions.problem
 
+import com.github.sharoha.cpjcodegen.actions.AddProblemAction
 import com.intellij.openapi.project.Project
 
 class LeetcodeProblemAction: AddProblemAction<LeetcodeProblemDialogWrapper>() {
     override fun getDialog(project: Project): LeetcodeProblemDialogWrapper {
-        TODO("Not yet implemented")
+        return LeetcodeProblemDialogWrapper(project)
     }
 
     override fun actionComplete() {
