+++ src/main/kotlin/com/github/sharoha/cpjcodegen/model/Platforms.kt
@@ -1,0 +1,5 @@
+package com.github.sharoha.cpjcodegen.model
+
+enum class Platforms(val platformName: String) {
+    LEETCODE("leetcode")
+}
