+++ src/main/resources/META-INF/plugin.xml
@@ -9,11 +9,11 @@
     <resource-bundle>messages.MyBundle</resource-bundle>
 
     <actions>
-        <action id="com.github.sharoha.cpjcodegen.actions.LeetcodeContestProblemAction"
-                class="com.github.sharoha.cpjcodegen.actions.LeetcodeContestProblemAction"
+        <action id="com.github.sharoha.cpjcodegen.actions.contest.LeetcodeContestProblemAction"
+                class="com.github.sharoha.cpjcodegen.actions.contest.LeetcodeContestProblemAction"
                 text="Leetcode Contest Parser" description="A task to parse leetcode current and previous contests"/>
-        <action id="com.github.sharoha.cpjcodegen.actions.LeetcodeProblemAction"
-                class="com.github.sharoha.cpjcodegen.actions.LeetcodeProblemAction" text="Leetcode Problem Parser"
-                description="An action to parse leetcode problem"/>
+<!--        <action id="com.github.sharoha.cpjcodegen.actions.problem.LeetcodeProblemAction"-->
+<!--                class="com.github.sharoha.cpjcodegen.actions.problem.LeetcodeProblemAction" text="Leetcode Problem Parser"-->
+<!--                description="An action to parse leetcode problem"/>-->
     </actions>
 </idea-plugin>
